pyproject.toml
[project]
name = "gst-billing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==4.2.2",
    "django-cors-headers==4.3.1",
    "django-crispy-forms==1.14.0",
    "django-htmx==1.12.0",
    "django-simple-history==3.2.0",
    "django-sql-explorer==3.0.1",
    "djangorestframework>=3.16.0",
    "ipython>=9.0.2",
    "num2words>=0.5.14",
    "openpyxl==3.0.10",
    "pre-commit>=4.1.0",
    "psycopg2-binary>=2.9.10",
    "ruff>=0.10.0",
    "setuptools>=76.0.0",
    "xlsxwriter==3.1.0",
]
